Silent compiler warning with MinGW64.

+++ src/scp.c
@@ -889,7 +889,7 @@ scp_send(LIBSSH2_SESSION * session, const char *path, int mode,
             session->scpSend_response_len =
                 snprintf((char *) session->scpSend_response,
                          LIBSSH2_SCP_RESPONSE_BUFLEN, "T%ld 0 %ld 0\n",
-                         mtime, atime);
+                         (long)mtime, (long)atime);
             _libssh2_debug(session, LIBSSH2_TRACE_SCP, "Sent %s",
                            session->scpSend_response);
         }
